@@ -5,7 +5,7 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^7.4|^8.0",
         "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^7.0.1",
@@ -21,8 +21,9 @@
         "mockery/mockery": "^1.4.2",
         "nunomaduro/collision": "^5.0",
         "nunomaduro/larastan": "^0.7.12",
-        "nunomaduro/phpinsights": "^1.0|^2.0",
+        "nunomaduro/phpinsights": "^2.0",
         "phpunit/phpunit": "^9.3.3",
+        "rector/rector": "^0.11.49",
         "squizlabs/php_codesniffer": "^3.6.0"
     },
     "autoload": {
