@@ -108,10 +108,9 @@ return [
         // MethodPerClassLimitSniff::class => [
         //     'maxCount' => 20,
         // ],
-        // COMPATIBILITY: Conditionally set this option for compatibility with PHPInsights:^1.0
-        FunctionLengthSniff::class => class_exists(FunctionLengthSniff::class, false) ? [
+        FunctionLengthSniff::class => [
             'maxLinesLength' => 25,
-        ] : null,
+        ],
     ]),
 
     /*
@@ -144,7 +143,6 @@ return [
     |
     */
 
-    // COMPATIBILITY: Unset this option for compatibility with PHPInsights:^1.0
-    // 'threads' => null,
+    'threads' => null,
 
 ];
